added `test` prefix to some tests
