+++ tests/_tests_scripts/test_core_wrapper.py
